% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing-utils.R
\name{parse_game_sqbs}
\alias{parse_game_sqbs}
\title{Parse game from an SQBS tournament}
\usage{
parse_game_sqbs(
  round,
  game_num,
  line,
  team1_name,
  team1,
  team2_name,
  team2,
  bonuses
)
}
\arguments{
\item{round}{The round of the tournament}

\item{game_num}{An integer indicating the game's order within the round on
the stats report}

\item{line}{The game score line}

\item{team1_name}{The name of the team listed first}

\item{team1}{The player score line of the team listed first}

\item{team2_name}{The name of the team listed second}

\item{team2}{The player score line of the team listed second}

\item{bonuses}{The bonus stat line}
}
\value{
Returns a named list of two data frames, one containing player stats
(conveniently named "player_stats") and the other team stats ("team_stats").
}
\description{
Parse game from an SQBS tournament
}
