% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing-utils.R
\name{parse_players_line_sqbs}
\alias{parse_players_line_sqbs}
\title{Parses player score lines from an SQBS tournament}
\usage{
parse_players_line_sqbs(player_line)
}
\arguments{
\item{player_line}{Line of player scores}

\item{powers}{Whether or not the tournament has powers ('T' or 'F')}
}
\value{
Returns a tidy data frame with columns for player, powers (if the tournament
has powers), tens, negs, and pts.
}
\description{
Parses player score lines from an SQBS tournament
}
