% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing-utils.R
\name{post_processing}
\alias{post_processing}
\title{Utility function to handle random stats issues}
\usage{
post_processing(df)
}
\arguments{
\item{df}{Arbitrary data frame}
}
\value{
Returns cleaned data frame
}
\description{
Utility function to handle random stats issues
}
