% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing-utils.R
\name{process_issues}
\alias{process_issues}
\title{Utility function to handle random stats issues}
\usage{
process_issues(html)
}
\arguments{
\item{html}{Arbitrary string}
}
\value{
Returns cleaned string
}
\description{
Utility function to handle random stats issues
}
