% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament-parser.R
\name{read_tournament_entry}
\alias{read_tournament_entry}
\title{Title}
\usage{
read_tournament_entry(year, set, site, url, powers, difficulty)
}
\arguments{
\item{year}{}

\item{set}{}

\item{site}{}

\item{url}{}

\item{powers}{}

\item{difficulty}{}
}
\value{

}
\description{
Title
}
